+++ warpgate-db-migrations/src/lib.rs
@@ -11,6 +11,7 @@ mod m00006_add_session_protocol;
 mod m00007_targets_and_roles;
 mod m00008_users;
 mod m00009_credential_models;
+mod m00010_parameters;
 
 pub struct Migrator;
 
@@ -27,6 +28,7 @@ impl MigratorTrait for Migrator {
             Box::new(m00007_targets_and_roles::Migration),
             Box::new(m00008_users::Migration),
             Box::new(m00009_credential_models::Migration),
+            Box::new(m00010_parameters::Migration),
         ]
     }
 }
